Monitor IDs
Monitor objects had their type id (the thing that used to tell the
engine what was in the monitor) replaced by an ID number that tells my
code which persistence slot it is.

Note that the code that *used* this change was already added, now we're
making it track individually instead of treating all blocks of a type as
if they were the same.