Inline line highlight
This mess of changes is for the sake of flipping the highlighted line
value inline instead of rerendering it.  Consequences though...

Moving the palette stuff into the drawing loop let me move the special
render to where the highlight was.  Only that is too long, so I cleaned
up the code to shrink it.  Best space saving was converting the text to
be pascal style strings, prefixed with length, which removed the
explicit line lengths but also allowed for the extra long lines to be
done in the loop.

With a few other slight tweaks, everything is properly aligned again.